requirements.txt
# --- FastAPI Core ---
fastapi==0.109.2
uvicorn==0.24.0.post1
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9

# --- Database ---
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9

# --- Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# --- Monitoring & Utils ---
prometheus-client==0.19.0
psutil==5.9.8
chromadb==0.4.24